\name{gs_gauge_list}
\alias{gs_gauge_list}
\title{List your gaug.es}
\usage{
  gs_gauge_list(keyname = "GaugesKey", page = NULL)
}
\arguments{
  \item{page}{Page to return.}

  \item{keyname}{Your API key name in your .Rprofile file}
}
\description{
  List your gaug.es
}
\examples{
\dontrun{
# scotts data
gs_gauge_list()
}
}

