% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gs_me}
\alias{gs_me}
\title{Information on yourself.}
\usage{
gs_me(keyname = "GaugesKey", callopts = list())
}
\arguments{
  \item{keyname}{Your API key name in your .Rprofile file}

  \item{callopts}{Curl debugging options passed in to
  httr::GET}
}
\description{
Information on yourself.
}
\note{
Note that these examples only work if you provide your
		Gaug.es API key either by passing in the function, or
		storing your key in your .Rprofile file, defaults to looking
		for an entry of "GaugesKey".
}
\examples{
\dontrun{
gs_me()
gs_me(callopts=verbose())
}
}

