% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gs_me.R
\name{gs_me_update}
\alias{gs_me_update}
\title{Updates and returns your information with the updates applied.}
\usage{
gs_me_update()
}
\description{
Updates and returns your information with the updates applied.
}

