% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauges.R
\name{parse_gauge_detail}
\alias{parse_gauge_detail}
\title{Parse guage detail}
\usage{
parse_gauge_detail(out)
}
\arguments{
\item{out}{Input from a gauge detail call.}
}
\description{
Parse guage detail
}
\keyword{internal}

