% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauges.R
\name{gs_gauge_list}
\alias{gs_gauge_list}
\title{List your gaug.es}
\usage{
gs_gauge_list(key = NULL, keyname = "GaugesKey", page = NULL, ...)
}
\arguments{
\item{key}{API key. If left NULL, function looks for key in your options settings
defined in the keyname parameter}

\item{keyname}{Your API key name in your .Rprofile file}

\item{page}{Page to return.}

\item{...}{Curl debugging options passed in to \code{\link[httr]{GET}}}
}
\description{
List your gaug.es
}
\note{
Note that these examples only work if you provide your
		Gaug.es API key either by passing in the function, or
		storing your key in your .Rprofile file, defaults to looking
		for an entry of "GaugesKey".
}
\examples{
\dontrun{
gs_gauge_list()
}
}

