% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gs_referrers.R
\name{gs_ref}
\alias{gs_ref}
\title{Gets referrers for a gauge, paginated.}
\usage{
gs_ref(id, date = NULL, page = NULL, key = NULL, keyname = "GaugesKey",
  ...)
}
\arguments{
\item{id}{Your gaug.es id}

\item{date}{Date format YYYY-MM-DD.}

\item{page}{page to return}

\item{key}{API key. If left NULL, function looks for key in your options settings
defined in the keyname parameter}

\item{keyname}{Your API key name in your .Rprofile file}

\item{...}{Curl debugging options passed in to \code{\link[httr]{GET}}}
}
\description{
Gets referrers for a gauge, paginated.
}
\note{
Note that these examples only work if you provide your
		Gaug.es API key either by passing in the function, or
		storing your key in your .Rprofile file, defaults to looking
		for an entry of "GaugesKey".
}
\examples{
\dontrun{
gs_ref(id='4efd83a6f5a1f5158a000004')

# Get list of gauge's, then pass in one of the ids
out <- gs_gauge_list()
gs_ref(id=out$brief[13,1])

library("httr")
gs_ref(id='4efd83a6f5a1f5158a000004', config=verbose())
}
}

