% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gs_tech}
\alias{gs_tech}
\title{Gets browsers and platforms for a gauge.}
\usage{
gs_tech(id, date = NULL, keyname = "GaugesKey", callopts = list())
}
\arguments{
  \item{callopts}{Curl debugging options passed in to
  httr::GET}

  \item{id}{Your gaug.es id}

  \item{date}{Date format YYYY-MM-DD. This works in a weird
  way. If you give no date, you get the traffic for each
  day since the beginning of the current month, but if you
  give a date, you get the traffic for each day for that
  entire month.}

  \item{keyname}{Your API key name in your .Rprofile file}
}
\value{
list of two, browsers and platforms
}
\description{
Gets browsers and platforms for a gauge.
}
\note{
Note that these examples only work if you provide your
		Gaug.es API key either by passing in the function, or
		storing your key in your .Rprofile file, defaults to looking
		for an entry of "GaugesKey".
}
\examples{
\dontrun{
gs_tech(id='4efd83a6f5a1f5158a000004')

# Get list of gauge's, then pass in one of the ids
out <- gs_gauge_list()
gs_tech(id=out$brief[6,1])
}
}

